Removed version check as it should be done on release number, tidied up colours thanks to Ikey's suggestion. Fixed README thanks to guitmz, cleaned up other minor bits including an if check before attempting to remove something

+++ ur
@@ -18,12 +18,24 @@ function actions {
     elif [ $1 == "update-repo" ] || [ $1 == "ur" ];then updaterepo $package
     elif [ $1 == "list-available" ] || [ $1 == "la" ]; then listpackages
     elif [ $1 == "list-installed" ] || [ $1 == "li" ]; then listinstalled
-    else echo -e "$red$error Incorrect syntax.";printhelp
+    else echo -e "${red}Error: ${white}Incorrect syntax.";printhelp
   fi
 }
 
+function addtoupgradelist {
+  pkgname=$1
+  release=$2
+  version=$3
+  oldrel=$4
+  oldver=$5
+  # Put package names in a list for it to process
+  echo $pkgname >> /tmp/ur/doup
+  # Create a CSV with package data in it
+  echo $pkgname,$version,$release,$oldver,$oldrel >> /tmp/ur/uplist
+}
+
 function createpackagerfile {
-  echo -e "$yellow$notice In order to build a package please enter the following:"
+  echo -e "${yellow}Notice: ${white}In order to build a package please enter the following:"
   mkdir -p ~/.solus
   touch ~/.solus/packager
   read -p "Full Name: " name
@@ -31,28 +43,28 @@ function createpackagerfile {
   echo -e "[Packager]" >> ~/.solus/packager
   echo -e "Name=$name" >> ~/.solus/packager
   echo -e "Email=$email"  >> ~/.solus/packager
-  echo -e "$yellow$notice Settings saved."
+  echo -e "${yellow}Notice: ${white}Settings saved."
 }
 
 function install {
   if [[ $package == "" ]]
     then
-      echo -e "$red$error No package name specified."
+      echo -e "${red}Error: ${white}No package name specified."
     else
       if [[ $(cat /usr/share/solus-user-repo/repo-index | grep $package | wc -l) -eq 0 ]];
         then
-          echo -e "$red$error Download failed or invalid package name specified."
+          echo -e "${red}Error: ${white}Download failed or invalid package name specified."
         else
           cd /tmp/ur
-          echo -e "$yellow$notice Attempting template download, this should only take a moment."
+          echo -e "${yellow}Notice: ${white}Attempting template download, this should only take a moment."
           wget -q http://solus-us.tk/ur/$1.yml
           mv $1.yml package.yml
-          echo -e "$yellow$notice Template found, building package."
+          echo -e "${yellow}Notice: ${white}Template found, building package."
           ypkg package.yml
-          if [[ $(find . -type f -iname "*.eopkg" | wc -l) -eq 0 ]];then echo -e "$red$error Build failed"
+          if [[ $(find . -type f -iname "*.eopkg" | wc -l) -eq 0 ]];then echo -e "${red}Error: ${white}Build failed"
             else
-                echo -e "$yellow$notice"
-                read -p "Build successful, install $package? (y/n) " -n 1 -r
+                echo -e "${yellow}Notice: ${white}Build of $package successful."
+                read -p "Install to your system? (y/n) " -n 1 -r
                 if [[ $REPLY =~ ^[Yy]$ ]]
                 then
                   # Do the install via eopkg
@@ -61,7 +73,7 @@ function install {
                 else
                   cp /tmp/ur/*.eopkg ~
                   echo ""
-                  echo -e "$yellow$notice Install aborted, eopkg file(s) are in your home directory."
+                  echo -e "${yellow}Notice: ${white}Install aborted, eopkg file(s) are in your home directory."
                 fi
           # Tell DB installed
           if [[ $(grep $package /usr/share/solus-user-repo/database | wc -l) -eq 0 ]];then
@@ -75,8 +87,8 @@ function install {
 }
 
 function listinstalled {
-  rm /tmp/ur/installed
-  echo -e "$yellow$notice Installed packages:"
+  if [[ -f /tmp/ur/installed ]];then rm /tmp/ur/installed;fi
+  echo -e "${yellow}Notice: ${white}Installed packages:"
   # Check what packages are installed from the database
   while read p; do
       if [[ $(echo $p | cut -d= -f 2) == 1 ]];
@@ -87,20 +99,20 @@ function listinstalled {
     # Check if any packages installed
     if [ ! -f /tmp/ur/installed ];
       then
-        echo -e "$yellow$notice Database shows no packages installed."
+        echo -e "${yellow}Notice: ${white}Database shows no packages installed."
       else
         cat /tmp/ur/installed
     fi
 }
 
 function listpackages {
-  echo -e "$yellow$notice Listing available packages."
+  echo -e "${yellow}Notice: ${white}Listing available packages."
   cat /usr/share/solus-user-repo/repo-index | more
 }
 
 function printhelp {
   echo -e ""
-  echo Usage:
+  echo "${yellow}Usage:${white}"
   echo -e "ur install (it) - Install a package (specify name)."
   echo -e "ur list-available (la) - List packages available in the user repository."
   echo -e "ur list-installed (li) - List packages installed from the user repository."
@@ -110,24 +122,26 @@ function printhelp {
   echo -e "ur upgrade (up) - Upgrade a package (specify name) or all (no value specified)."
   echo -e "ur viewinfo (vi) - View information on a package in the repository (specify name)"
   echo -e ""
-  echo -e "Examples:"
+  echo -e "${yellow}Examples:${white}"
   echo -e "ur install dfc"
   echo -e "ur up pantheon-photos"
 }
 
 function remove {
-  if [[ $package == "" ]];then echo -e "$red$error No package name specified."
+  if [[ $package == "" ]];then echo -e "${red}Error: ${white}No package name specified."
     else
       if [[ $confirm == "-y" ]]
         then
-          echo -e "$yellow$notice Removing $package from your system."
+          echo -e "${yellow}Notice: ${white}Removing $package from your system."
           sudo eopkg rm $package
           sed -i 's/'"$package"'=1/'"$package"'=0/g' /usr/share/solus-user-repo/database
         else
-          read -p "$yellow$notice Do you wish to remove $package? (y/n) " -n 1 -r
+          echo -e "${yellow}Notice: ${white}Do you wish to remove $package?"
+          read -p "Confirm (y/n) " -n 1 -r
             if [[ $REPLY =~ ^[Yy]$ ]]
               then
-                echo -e "$yellow$notice Removing $package from your system."
+                echo ""
+                echo -e "${yellow}Notice: ${white}Removing $package from your system."
                 sudo eopkg rm $package
                 sed -i 's/'"$package"'=1/'"$package"'=0/g' /usr/share/solus-user-repo/database
             fi
@@ -136,12 +150,12 @@ function remove {
 }
 
 function search {
-  if [[ $package == "" ]];then echo -e "$red$error No search term provided."
+  if [[ $package == "" ]];then echo -e "${red}Error: ${white}No search term provided."
   else
-  echo -e "Searching for $package."
-    if [[ $(grep -i $package /usr/share/solus-user-repo/repo-index | wc -l) -eq 0 ]];then echo -e "$red$error No results for $package."
-      elif [[ $(grep -i $package /usr/share/solus-user-repo/repo-index | wc -l) -gt 1 ]];then echo -e "$yellow$notice Found $(grep -i $package /usr/share/solus-user-repo/repo-index | wc -l) items:";grep -i $package /usr/share/solus-user-repo/repo-index
-      else echo -e "$yellow$notice Found 1 item:";grep -i $package /usr/share/solus-user-repo/repo-index
+  echo -e "${yellow}Notice: ${white}Searching for $package."
+    if [[ $(grep -i $package /usr/share/solus-user-repo/repo-index | wc -l) -eq 0 ]];then echo -e "${red}Error: ${white}No results for $package."
+      elif [[ $(grep -i $package /usr/share/solus-user-repo/repo-index | wc -l) -gt 1 ]];then echo -e "${yellow}Notice: ${white}Found $(grep -i $package /usr/share/solus-user-repo/repo-index | wc -l) items:";grep -i $package /usr/share/solus-user-repo/repo-index
+      else echo -e "${yellow}Notice: ${white}Found 1 item:";grep -i $package /usr/share/solus-user-repo/repo-index
     fi
   fi
 }
@@ -154,7 +168,7 @@ function updaterepo {
 function upgrade {
   # Check if we're upgrading all or one specific package, or all with skipyn
   if [[ $package == "-y" ]] || [[ $package == "" ]];then
-    echo -e "$yellow$notice Checking what packages need upgrading."
+    echo -e "${yellow}Notice: ${white}Checking what packages need upgrading."
     # Check what packages are installed
     while read p; do
       if [[ $(echo $p | cut -d= -f 2) == 1 ]];
@@ -165,22 +179,34 @@ function upgrade {
     # Check if anything found
     if [ -f /tmp/ur/upgrades ];
       then
-        # Do version checks against packages
+        # Do release number checks against packages
         while read a; do
           cd /tmp/ur
           wget -q http://solus-us.tk/ur/$a.yml
-          newver=$(cat $a.yml | grep version | cut -d: -f 2 | sed 's/ //g')
-          inver=$(eopkg info $a | grep Name | cut -d: -f 3 | cut -d, -f 1 | sed 's/ //g')
-          if [[ $(vercomp $inver $newver) -eq 1 ]];then echo $a >> /tmp/ur/doup
+          reporelease=$(cat $a.yml | grep release | cut -d: -f 2 | sed 's/ //g')
+          repoversion=$(cat $a.yml | grep version | cut -d: -f 2 | sed 's/ //g')
+          installedrelease=$(eopkg info $a | grep Name | cut -d: -f 4 | sed 's/ //g')
+          installedversion=$(eopkg info $a | grep Name | cut -d: -f 3 | cut -d, -f 1 | sed 's/ //g')
+          if [[ $reporelease -gt $installedrelease ]]
+            then addtoupgradelist $a $reporelease $repoversion $installedrelease $installedversion
           fi
           echo 
         done </tmp/ur/upgrades
-        echo -e "$yellow$notice Upgrade checks done."
+        echo -e "${yellow}Notice: ${white}Upgrade checks done."
         # Check if any were found that were higher version
         if [ -f /tmp/ur/doup ];
           then
-            echo -e "$yellow$notice The following packages will be upgraded:"
-            cat /tmp/ur/doup | sort
+            echo -e "${yellow}Notice: ${white}The following package(s) will be upgraded:"
+            # Do a fancy upgrade list here
+            while read m; do
+              #do stuff to $m
+              pkg=$(echo $m | cut -d, -f 1)
+              ver=$(echo $m | cut -d, -f 2)
+              rel=$(echo $m | cut -d, -f 3)
+              oldver=$(echo $m | cut -d, -f 4)
+              oldrel=$(echo $m | cut -d, -f 5)
+              echo -e "${yellow}Package: ${white}$pkg ${yellow}Version: ${white}$oldver ${yellow}Release: ${white}$oldrel to be upgraded to ${yellow}Version: ${white}$ver ${yellow}Release: ${white}$rel"
+            done </tmp/ur/urlist
             # Check for skipyn
             if [[ $package == "-y" ]];then
               # Do the upgrades
@@ -198,10 +224,10 @@ function upgrade {
             fi
           fi
           else
-            echo -e "$yellow$notice No packages need upgrading."
+            echo -e "${yellow}Notice: ${white}No packages need upgrading."
         fi
       else
-        echo -e "$yellow$notice No packages need upgrading."
+        echo -e "${yellow}Notice: ${white}No packages need upgrading."
     fi
   else
     # Single package upgrade, package name specified
@@ -212,7 +238,8 @@ function upgrade {
     if [[ $skipyn == "1" ]];
       then installpackage $package
     else
-      read -p "$yellow$notice Do you wish to proceed? (y/n) " -n 1 -r
+      echo -e "${yellow}Notice: ${white}Package built."
+      read -p "Do you wish to proceed? (y/n) " -n 1 -r
         if [[ $REPLY =~ ^[Yy]$ ]]
           then
             installpackage $package
@@ -221,44 +248,12 @@ function upgrade {
 fi
 }
 
-# Taken from http://stackoverflow.com/questions/4023830/bash-how-compare-two-strings-in-version-format by Dennis Williamson http://stackoverflow.com/users/26428/dennis-williamson
-function vercomp () {
-    if [[ $1 == $2 ]]
-    then
-        return 0
-    fi
-    local IFS=.
-    local i ver1=($1) ver2=($2)
-    # fill empty fields in ver1 with zeros
-    for ((i=${#ver1[@]}; i<${#ver2[@]}; i++))
-    do
-        ver1[i]=0
-    done
-    for ((i=0; i<${#ver1[@]}; i++))
-    do
-        if [[ -z ${ver2[i]} ]]
-        then
-            # fill empty fields in ver2 with zeros
-            ver2[i]=0
-        fi
-        if ((10#${ver1[i]} > 10#${ver2[i]}))
-        then
-            return 1
-        fi
-        if ((10#${ver1[i]} < 10#${ver2[i]}))
-        then
-            return 2
-        fi
-    done
-    return 0
-}
-
 function viewpackage {
   if [[ $package == "" ]]
     then
-      echo -e "$red$error No package name supplied."
+      echo -e "${red}Error: ${white}No package name supplied."
     else
-      echo -e "$yellow$notice Getting package info for $package"
+      echo -e "${yellow}Notice: ${white}Getting package info for $package"
       cd /tmp/ur
       wget -q http://solus-us.tk/ur/$package.yml
       cat $package.yml
@@ -270,12 +265,9 @@ function viewpackage {
 package=$2
 action=$1
 confirm=$3
-warning="Warning:"
-error="Error:$white"
-notice="Notice:$white"
 
 # Warning message
-echo -e "$red$warning$white"
+echo -e "${red}Warning:${white}"
 echo -e "* All items in the Solus User Repo are used at your own risk."
 echo -e "* Templates are created by Solus users and have NOT been tested."
 
@@ -290,7 +282,7 @@ if [[ ! -d ~/.solus ]];then createpackagerfile
 fi
 
 # Check if repo index exists
-if [[ ! -f /usr/share/solus-user-repo/repo-index ]];then echo -e "$yellow$notice Repository index not present, fetching.";updaterepo
+if [[ ! -f /usr/share/solus-user-repo/repo-index ]];then echo -e "${yellow}Notice: ${white}Repository index not present, fetching.";updaterepo
 fi
 
 # Check if database exists if not create
